+++ src/view/App.tsx
@@ -29,7 +29,7 @@ import { any, AppRoute } from "./routes";
 import { Header } from "./screen/home/Header";
 
 import { CreatePassword, Initialize } from "./screen/initialize/Initialize";
-import { LadgerNotification } from "./screen/ladger/LadgerNotification";
+import { LedgerNotification } from "./screen/ledger/LedgerNotification";
 import { Loading } from "./screen/Loading";
 import { Unlock } from "./screen/unlock/Unlock";
 import { WebAuthnNotification } from "./screen/unlock/WebAuthnNotification";
@@ -95,7 +95,7 @@ const ContentRouter: FC<{
             <Route path="*" element={<Home />} />
           </Routes>
         </Suspense>
-        <LadgerNotification />
+        <LedgerNotification />
       </WalletContractContext.Provider>
     </WalletStateContext.Provider>
   );