+++ src/view/screen/import/hardware/api.tsx
@@ -2,20 +2,20 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useContext } from "react";
 import { TonTransport } from "ton-ledger";
 import { WalletState } from "../../../../libs/entries/wallet";
-import { ladgerPathForAccount } from "../../../../libs/service/transfer/ladger";
+import { LedgerPathForAccount } from "../../../../libs/service/transfer/ledger";
 import { AccountStateContext, NetworkContext } from "../../../context";
 import { saveAccountState } from "../../api";
 
 let workchain = 0;
 let chain = workchain === -1 ? 255 : 0;
 
-export const getLadgerWalletState = async (
+export const getLedgerWalletState = async (
   network: string,
   transport: TonTransport,
   accountIndex: number
 ): Promise<WalletState> => {
   let bounceable = false;
-  let path = ladgerPathForAccount(network, workchain, accountIndex);
+  let path = LedgerPathForAccount(network, workchain, accountIndex);
   let response = await transport.getAddress(path, {
     chain,
     bounceable,
@@ -25,24 +25,24 @@ export const getLadgerWalletState = async (
   let address: string = response.address;
 
   return {
-    name: `Ladger ${accountIndex + 1}`,
+    name: `Ledger ${accountIndex + 1}`,
     mnemonic: "",
     address,
     publicKey: publiKey.toString("hex"),
     version: "v4R2",
     isBounceable: bounceable,
-    isLadger: true,
-    ladgerIndex: accountIndex,
-    ladgerDriver: "USB",
+    isLedger: true,
+    LedgerIndex: accountIndex,
+    LedgerDriver: "USB",
   };
 };
 
-export const useLadgerAccounts = () => {
+export const useLedgerAccounts = () => {
   const network = useContext(NetworkContext);
   return useMutation<WalletState[], Error, TonTransport>(async (transport) => {
     return Promise.all(
       [0, 1, 2, 3, 4].map((index) =>
-        getLadgerWalletState(network, transport, index)
+        getLedgerWalletState(network, transport, index)
       )
     );
   });